Reloading Animations, Changes to reloading functionality, muzzle flash
Setup reloading animation and reloading functionality to work after the animation finishes. Mainly done in blueprints but tweaked the reload logic to use a boolean in code that we expose to blueprints and disable once the animation has finished. Also forced the followcamera to be used while the reload animation is happening and made sure we can't fire while reloading.

Also setup muzzle flash particle effect to spawn at our muzzle socket when firing.